move Axis to benchmarks_view
